[Move]Move out from definitions.go and help.go
